fix root-finding: use cwd instead of location of versioneer.py
This fixes the case where a setuptools-based 'install' or 'develop' is
used on a project which uses Versioneer, which also depends upon another
project that uses Versioneer. These invocations work by running each
dependency's setup.py *in the same process*, so when the dependencies do
"import versioneer", they find the top-level project's versioneer.py,
and then look at the wrong setup.cfg.

By using the current working directory as a root, we find the right
setup.cfg . This does not help the case where the dependencies are using
a different *version* or versioneer.py: that must be addressed by
something else.