override 'build_py', not 'build': fix setuptools install/bdist_egg
This should fix the setuptools 'install' and 'bdist_egg' commands, which
call 'build_py' but not 'build'. Previously, by overriding 'build', we
were not getting control during these commands, leaving an unmodified
_version.py in the repo, which then failed to find its version,
resulting in a "0+unknown" string.

The total graph of invocation pathways is now:

 bdist_wheel -> distutils/install -> distutils/build -> distutils/build_py
 setuptools/install -> bdist_egg -> distutils/install_lib -> build_py
 distutils/sdist
 setuptools/sdist
 setuptools/develop: unnecessary, imports source from VCS repo

We do not yet correctly handle setuptools/sdist: this version upcalls to
the distutils class, so it ignores anything special done by setuptools'
version of the 'sdist' command (mainly adding the .egg-info directory).
A later patch will fix that.