cmdclass: handle setuptools 'sdist' properly
Previously, our 'sdist' override was upcalling to the distutils version
even if we were run in setuptools environment. The symptom was that the
sdist tarballs were lacking the expected .egg-info directory.

This approach looks in sys.modules to determine whether setuptools has
been imported or not (setuptools monkeypatches distutils' Distribution
class upon import), and upcalls to the right parent class accordingly.